test: Re-add basic test for cli module
Readding tests for the cli module. Previous work was merged after
the record_history changes, causing failures in main. This should
resolve.

As with the original work, this does not provide much cli coverage
yet, but given the recent breakage I figured it was worth getting a
start of cli tests.